Ya se muestra la pantalla de publicación, y se muestran comentarios
Da error al agregar comentarios pero si se agregan, falta responder comentarios

+++ PlayPostAccesoADatos/src/main/java/facade/FacadeComentario.java
@@ -101,7 +101,7 @@ public class FacadeComentario implements IFacadeComentario {
      * encuentra.
      */
     @Override
-    public Comentario buscarComentarioPorID(int id) {
+    public Comentario buscarComentarioPorID(Long id) {
         try {
             return factory.crearDAO().buscarComentarioPorID(id);
         } catch (PersistenciaException ex) {
